Revert blog section to previous state
