+++ app/blog/[singlePost]/page.js
@@ -1,54 +1,40 @@
-import { notFound } from "next/navigation";
+"use client";
 
-import SinglePostPageContent from "@/components/blog/SinglePostPageContent";
-import posts from "@/lib/posts.json";
+import { useRouter } from "next/navigation";
+import { useCategory } from "@/components/blog/CategoryContext";
 
-const siteTitle = "2024 體驗營切版任務二 by Aaron";
+import styles from "./SinglePostPage.module.css";
 
-const normalizeSlug = (slug = "") =>
-  decodeURIComponent(slug).replace(/^\/+/, "").replace(/\/+$/, "");
+import SinglePost from "@/components/blog/Posts/SinglePost";
+import BlogMobileNav from "@/components/blog/BlogMobileNav";
+import BlogSideBar from "@/components/blog/BlogSideBar";
+import RelatedPosts from "@/components/blog/Posts/RelatedPosts";
 
-export function generateStaticParams() {
-  return posts.map((post) => ({ singlePost: normalizeSlug(post.postSlug) }));
-}
+export default function SinglePostPage() {
+  const { activeCategory, handleCategoryChange } = useCategory();
+  const router = useRouter();
 
-export function generateMetadata({ params }) {
-  const slug = normalizeSlug(params.singlePost);
-  const post = posts.find((item) => normalizeSlug(item.postSlug) === slug);
-
-  if (!post) {
-    return {
-      title: `部落格 | ${siteTitle}`,
-      description: "找不到指定的文章。",
-    };
-  }
-
-  return {
-    title: `${post.postMeta.title} | ${siteTitle}`,
-    description: post.postMeta.summary,
-    openGraph: {
-      title: `${post.postMeta.title} | ${siteTitle}`,
-      description: post.postMeta.summary,
-      url: `https://hex2024.worksbyaaron.com/blog/${params.singlePost}`,
-      siteName: siteTitle,
-      locale: "zh_TW",
-      type: "article",
-    },
-    twitter: {
-      title: `${post.postMeta.title} | ${siteTitle}`,
-      description: post.postMeta.summary,
-      card: "summary_large_image",
-    },
+  const handleCategoryClick = (category) => {
+    handleCategoryChange(category);
+    router.push("/blog", undefined, { shallow: true });
   };
-}
 
-export default function SinglePostPage({ params }) {
-  const slug = normalizeSlug(params.singlePost);
-  const post = posts.find((item) => normalizeSlug(item.postSlug) === slug);
-
-  if (!post) {
-    notFound();
-  }
-
-  return <SinglePostPageContent post={post} />;
+  return (
+    <>
+      <BlogMobileNav
+        activeCategory={activeCategory}
+        onCategoryChange={handleCategoryClick}
+      />
+      <section className={styles.container}>
+        <article>
+          <SinglePost />
+        </article>
+        <BlogSideBar
+          activeCategory={activeCategory}
+          onCategoryChange={handleCategoryClick}
+        />
+      </section>
+      <RelatedPosts />
+    </>
+  );
 }
