+++ app/blog/page.js
@@ -1,11 +1,25 @@
-import BlogPage from "@/components/blog/BlogPage";
+"use client";
 
-export const metadata = {
-  title: "部落格",
-  description:
-    "瀏覽最新的 UI/UX 觀點與產品設計心得，掌握 2024 體驗營切版任務二 的創作歷程。",
-};
+import { useCategory } from "@/components/blog/CategoryContext";
+
+import BlogHeading from "@/components/blog/BlogHeading";
+import BlogMobileNav from "@/components/blog/BlogMobileNav";
+import BlogPosts from "@/components/blog/BlogPost";
 
 export default function Blog() {
-  return <BlogPage />;
+  const { activeCategory, handleCategoryChange } = useCategory();
+
+  return (
+    <>
+      <BlogHeading />
+      <BlogMobileNav
+        activeCategory={activeCategory}
+        onCategoryChange={handleCategoryChange}
+      />
+      <BlogPosts
+        activeCategory={activeCategory}
+        onCategoryChange={handleCategoryChange}
+      />
+    </>
+  );
 }
