+++ components/blog/BlogHeading.js
@@ -3,7 +3,6 @@ import styles from "./BlogHeading.module.css";
 import HeadingWithLine from "../ui/HeadingWithLine";
 
 export default function BlogHeading() {
-  // 部落格頁面標題區，凸顯主要標題與說明文字
   return (
     <section className={styles.container}>
       <HeadingWithLine headingContent="部落格" isH1 />
