--- components/blog/BlogPage.js
@@ -1,25 +1,0 @@
-"use client";
-
-import { useCategory } from "@/components/blog/CategoryContext";
-
-import BlogHeading from "@/components/blog/BlogHeading";
-import BlogMobileNav from "@/components/blog/BlogMobileNav";
-import BlogPosts from "@/components/blog/BlogPost";
-
-export default function BlogPage() {
-  const { activeCategory, handleCategoryChange } = useCategory();
-
-  return (
-    <>
-      <BlogHeading />
-      <BlogMobileNav
-        activeCategory={activeCategory}
-        onCategoryChange={handleCategoryChange}
-      />
-      <BlogPosts
-        activeCategory={activeCategory}
-        onCategoryChange={handleCategoryChange}
-      />
-    </>
-  );
-}
